requirements.txt
aiobotocore==2.24.2
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aioitertools==0.12.0
aiosignal==1.4.0
attrs==25.3.0
blinker==1.9.0
boto3==1.40.28
botocore==1.40.18
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
Flask==3.1.2
frozenlist==1.7.0
fsspec==2025.9.0
greenlet==3.2.4
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
kagglehub==0.3.13
MarkupSafe==3.0.2
multidict==6.6.4
narwhals==2.5.0
numpy==2.3.3
packaging==25.0
pandas==2.3.2
plotly==6.3.0
propcache==0.3.2
psycopg2-binary==2.9.10
pyarrow==21.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.5
s3fs==2025.9.0
s3transfer==0.14.0
six==1.17.0
SQLAlchemy==2.0.43
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
wrapt==1.17.3
yarl==1.20.1
